Revert "Merge remote-tracking branch 'upstream/item_expansion' into E-H"
This reverts commit fdf1f8c92f28025a074e5a712ac0d75ed9e2efe9, reversing
changes made to 4b2155999568ff55bb923f332c42bc8e44e0f186.

+++ src/pokenav_unk_4.c
@@ -126,7 +126,7 @@ bool32 sub_81CB260(void)
     return TRUE;
 }
 
-void sub_81CB29C(s32 index)
+void sub_81CB29C(int index)
 {
     struct Pokenav4Struct *state = GetSubstructPtr(6);
     state->unk4 = CreateLoopedTask(gUnknown_08622798[index], 1);
@@ -706,7 +706,7 @@ u32 sub_81CBB74(int taskState)
 static void sub_81CBBB8(void)
 {
     struct MatchCallListTemplate template;
-    template.unk0 = sub_81CAE94();
+    template.unk0 = (u32)sub_81CAE94();
     template.unk4 = sub_81CAE38();
     template.unk8 = 4;
     template.unk6 = 0;
